rename prompt to generate
