+++ cli/package.json
@@ -70,7 +70,7 @@
     "test:watch": "jest --watch",
     "ts-node-esm": "node --loader ts-node/esm",
     "transpile": "node --loader ts-node/esm src/index.ts transpile",
-    "prompt": "node --loader ts-node/esm/transpile-only src/index.ts prompt",
+    "generate": "node --loader ts-node/esm/transpile-only src/index.ts generate",
     "encode": "./node_modules/.bin/env-cmd -f .env ./node_modules/.bin/bitloops-gherkin encode -t",
     "package": "pkg ./dist/bitloops-language.bundle.js --targets node16-macos-x64 --output ./dist/bitloops-language --assets *.flf",
     "caxa": "npx caxa --input \"/Users/vasilis/Code/Bitloops/bitloops-language-cli\"  --output \"/Users/vasilis/Code/Bitloops/bitloops-language-cli/dist/Bitloops Language.app\" -- \"{{caxa}}/node_modules/.bin/node\" \"{{caxa}}/lib/src/index.js\" \"transpile\"",
