+++ cli/src/commands/generate/client.ts
