+++ cli/src/commands/generate/commands/code-davinci-command.ts
