+++ cli/src/commands/generate/commands/command.ts
