+++ cli/src/commands/generate/commands/gpt4-command.ts
