+++ cli/src/commands/generate/data-sets/common/names.ts
