+++ cli/src/commands/generate/invoker.ts
