+++ cli/src/commands/generate/static-assets/apiModule.ts
