+++ cli/src/commands/generate/static-assets/authenticationController.ts
