+++ cli/src/commands/generate/static-assets/config.ts
